+++ src/candidate-packages/authorize-flow/AuthModal.tsx
@@ -34,22 +34,27 @@ export const AuthModal: React.FC<AuthRedirectModalProps> = ({ signOutUrl, deboun
     return unsubscribe;
   }, []);
 
+  const handleCloseClick = () => {
+    setIsOpen(false);
+  };
+
   return (
-    <Modal hideCloseButton onClose={() => { }} sx={{ m: 2, p: 2 }} open={isOpen}>
+    <Modal hideCloseButton onClose={() => {}} sx={{ m: 2, p: 2 }} open={isOpen}>
       <FlexBox sx={{ p: 2, overflowY: "auto" }}>
-        <H3>
-          <i className="fa-regular fa-circle-exclamation"></i> Your session is no longer active
-        </H3>
+        <H3>Your session is no longer active</H3>
         <Text sx={{ pt: 4 }}>
           Your session may have expired, you may have signed out, or your account isn&apos;t active.
         </Text>
         <Text>Please login again to continue.</Text>
         <Text sx={{ pt: 4 }}>If you continue to have issues, contact your system administrator.</Text>
-        <Box ml="auto" mt={2}>
+        <FlexBox ml="auto" mt={2} gap={1} direction="row">
+          <Button variant="secondary" disableElevation onClick={handleCloseClick} size="large">
+            Close
+          </Button>
           <Button variant="primary" onClick={handleLoginClick} size="large">
             Login
           </Button>
-        </Box>
+        </FlexBox>
       </FlexBox>
     </Modal>
   );