+++ src/v1/theme/createTheme.test.ts
@@ -16,10 +16,7 @@ test("tmp theme diffs via unified patches", () => {
     isDarkSet.map((isDark) => {
       asMock(mockCreateMuiTheme).mockClear();
       createTheme(name, isDark, true);
-      return [
-        `${name} (${isDark ? "light" : "dark"})`,
-        asMock(mockCreateMuiTheme).mock.calls[0][0],
-      ];
+      return [`${name} (${isDark ? "light" : "dark"})`, asMock(mockCreateMuiTheme).mock.calls[0][0]];
     })
   );
 
@@ -28,13 +25,7 @@ test("tmp theme diffs via unified patches", () => {
   const report = pairs
     .slice(1)
     .map(([name, opts], idx) => {
-      const patch = createPatch(
-        `${idx}`,
-        baseOptsStr,
-        JSON.stringify(opts, null, 2),
-        baseName,
-        name
-      );
+      const patch = createPatch(`${idx}`, baseOptsStr, JSON.stringify(opts, null, 2), baseName, name);
       return patch;
     })
     .join("\n\n");
@@ -381,12 +372,12 @@ test("tmp theme diffs via unified patches", () => {
                  "borderWidth": 1,
                  "borderStyle": "solid",
     -            "borderColor": "rgba(141, 153, 226, 0.4)"
-    +            "borderColor": "inherit"
+    +            "borderColor": "rgba(0, 0, 0, 0.4)"
                }
              }
            ],
            "styleOverrides": {}
-    @@ -80,12 +80,12 @@
+    @@ -77,11 +77,11 @@
        },
        "palette": {
          "mode": "dark",
@@ -394,15 +385,13 @@ test("tmp theme diffs via unified patches", () => {
     -      "main": "#2F44CA",
     -      "light": "#5969D5",
     -      "dark": "#2636A2",
-    -      "contrastText": "#fff"
-    +      "main": "#fff",
-    +      "light": "rgba(255, 255, 255, 0.5)",
-    +      "dark": "rgba(255, 255, 255, 0.9)",
-    +      "contrastText": "#000"
+    +      "main": "#000000",
+    +      "light": "rgba(0, 0, 0, 0.5)",
+    +      "dark": "rgba(0, 0, 0, 0.7)",
+           "contrastText": "#fff"
          },
          "secondary": {
            "main": "#222222"
-         },
 
 
     Index: 8
@@ -415,7 +404,7 @@ test("tmp theme diffs via unified patches", () => {
                  "borderWidth": 1,
                  "borderStyle": "solid",
     -            "borderColor": "rgba(141, 153, 226, 0.4)"
-    +            "borderColor": "inherit"
+    +            "borderColor": "rgba(0, 0, 0, 0.4)"
                }
              }
            ],
@@ -431,11 +420,10 @@ test("tmp theme diffs via unified patches", () => {
     -      "main": "#2F44CA",
     -      "light": "#5969D5",
     -      "dark": "#2636A2",
-    -      "contrastText": "#fff"
-    +      "main": "#fff",
-    +      "light": "rgba(255, 255, 255, 0.5)",
-    +      "dark": "rgba(255, 255, 255, 0.9)",
-    +      "contrastText": "#000"
+    +      "main": "#000000",
+    +      "light": "rgba(0, 0, 0, 0.5)",
+    +      "dark": "rgba(0, 0, 0, 0.7)",
+           "contrastText": "#fff"
          },
          "secondary": {
     -      "main": "#222222"